Move the normal case to the front side
It is the reverse of my previous (and the standard library) style, but
is adapted to the optimization of ocamlopt that expects true cases to be
more likely.

However, this kind of small optimization is meaningless for this library
since the subject is I/O.
It's just a style fix.